keyrock added

+++ install.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 
 #
 # Ramp Installation Script
@@ -12,27 +12,35 @@ IFS=$(printf '\n\t')
 sudo mkdir /etc/grafana
 sudo cp -r provisioning /etc/grafana/
 # Docker
-sudo apt remove --yes docker docker-engine docker.io containerd runc
-sudo apt update
-sudo apt --yes --no-install-recommends install apt-transport-https ca-certificates
-wget --quiet --output-document=- https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo add-apt-repository "deb [arch=$(dpkg --print-architecture)] https://download.docker.com/linux/ubuntu $(lsb_release --codename --short) stable"
-sudo apt update
-sudo apt --yes --no-install-recommends install docker-ce docker-ce-cli containerd.io
-sudo usermod --append --groups docker "$USER"
-sudo systemctl enable docker
-printf '\nDocker installed successfully\n\n'
-printf 'Waiting for Docker to start...\n\n'
-sleep 5
+if [[ $(docker -v) == *" 20.10."* ]]; then
+    echo 'Docker version sufficient'
+else
+    echo 'Docker update needed'
+	sudo apt remove --yes docker docker-engine docker.io containerd runc
+	sudo apt update
+	sudo apt --yes --no-install-recommends install apt-transport-https ca-certificates
+	wget --quiet --output-document=- https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+	sudo add-apt-repository "deb [arch=$(dpkg --print-architecture)] https://download.docker.com/linux/ubuntu $(lsb_release --codename --short) stable"
+	sudo apt update
+	sudo apt --yes --no-install-recommends install docker-ce docker-ce-cli containerd.io
+	sudo usermod --append --groups docker "$USER"
+	sudo systemctl enable docker
+	printf '\nDocker installed successfully\n\n'
+	printf 'Waiting for Docker to start...\n\n'
+	sleep 5
+fi
 
 # Docker Compose
-sudo wget --output-document=/usr/local/bin/docker-compose "https://github.com/docker/compose/releases/download/$(wget --quiet --output-document=- https://api.github.com/repos/docker/compose/releases/latest | grep --perl-regexp --only-matching '"tag_name": "\K.*?(?=")')/run.sh"
-sudo chmod +x /usr/local/bin/docker-compose
-sudo wget --output-document=/etc/bash_completion.d/docker-compose "https://raw.githubusercontent.com/docker/compose/$(docker-compose version --short)/contrib/completion/bash/docker-compose"
-printf '\nDocker Compose installed successfully\n\n'
+if [[ $(docker-compose version) == *" 1.29."* ]]; then
+    echo 'Docker version sufficient'
+else
+	sudo wget --output-document=/usr/local/bin/docker-compose "https://github.com/docker/compose/releases/download/$(wget --quiet --output-document=- https://api.github.com/repos/docker/compose/releases/latest | grep --perl-regexp --only-matching '"tag_name": "\K.*?(?=")')/run.sh"
+	sudo chmod +x /usr/local/bin/docker-compose
+	sudo wget --output-document=/etc/bash_completion.d/docker-compose "https://raw.githubusercontent.com/docker/compose/$(docker-compose version --short)/contrib/completion/bash/docker-compose"
+	printf '\nDocker Compose installed successfully\n\n'
 
-sleep 5
-
+	sleep 5
+fi
 #
 # Copy docker daemon conf file
 #
